chore: refactor bookmark file locations
