+++ src/Core/WorkSpace/src/Context/BookmarkContext.tsx
@@ -1,51 +1,24 @@
 import { deref } from '@dbeining/react-atom';
-import {
-    ApplyEventArgs,
-    SaveArgs,
-    SaveEventArgs,
-    useBookmarkEvents,
-    useBookmarks,
-} from '@equinor/BookmarksManager';
-import { FilterGroup, useFilterApiContext } from '@equinor/filter';
+import { ApplyEventArgs, SaveEventArgs, useBookmarks } from '@equinor/BookmarksManager';
+import { useFilterApiContext } from '@equinor/filter';
 import { PowerBIBookmarkPayload } from '@equinor/lighthouse-powerbi';
 import { createContext, PropsWithChildren, useContext } from 'react';
-import { gardenApiAtom } from '../Util/gardenBookmarks/gardenApiAtom';
-import { WorkspaceTab } from '../WorkSpaceApi/workspaceState';
+import {
+    ApplyBookmark,
+    GardenPayload,
+    SaveBookmark,
+    WorkspaceBookmarkPayload,
+} from '../Util/bookmarks/types';
 import { useLocationContext } from './LocationProvider';
 import { useViewerContext } from './ViewProvider';
-type GardenPayload = {
-    customGroupByKeys?: Record<string, unknown>;
-    gardenKey?: string;
-    groupByKeys?: string[];
-};
-export type WorkspaceBookmarkPayload = {
-    activeTab: WorkspaceTab;
-    garden: GardenPayload;
-    filter: FilterGroup[];
-};
-type SavePBIWorkspaceBookmark = (saveEventArgs: SaveArgs<PowerBIBookmarkPayload>) => Promise<void>;
-type SaveWorkspaceBookmark = (saveEventArgs: SaveEventArgs) => Promise<void>;
-
-type SaveBookmark<T> = T extends PowerBIBookmarkPayload
-    ? SavePBIWorkspaceBookmark
-    : T extends WorkspaceBookmarkPayload
-    ? SaveWorkspaceBookmark
-    : unknown;
-type ApplyBookmark<T> = T extends PowerBIBookmarkPayload
-    ? PowerBIBookmarkPayload | undefined
-    : T extends WorkspaceBookmarkPayload
-    ? void
-    : unknown;
+import { isWorkspaceBookmark } from '../Util/bookmarks/helpers';
+import { gardenApiAtom } from '../Util/bookmarks/gardenBookmarks';
 
 type Context<T> = {
     applyBookmark: (args: ApplyEventArgs) => Promise<ApplyBookmark<T>>;
     saveBookmark: SaveBookmark<T>;
 };
-const isWorkspaceBookmark = (
-    bookmark: Awaited<WorkspaceBookmarkPayload | PowerBIBookmarkPayload | never>
-): bookmark is WorkspaceBookmarkPayload => {
-    return (bookmark as WorkspaceBookmarkPayload)?.activeTab !== undefined;
-};
+
 const BookmarkContext = createContext({} as Context<unknown>);
 
 type BookmarkContextWrapperProps = {};
@@ -153,10 +126,4 @@ export const useBookmarkContext = <
     T extends PowerBIBookmarkPayload | WorkspaceBookmarkPayload
 >(): Context<T> => {
     return useContext(BookmarkContext);
-};
-
-export const useWorkspaceBookmarks = () => {
-    const { applyBookmark, saveBookmark } = useBookmarkContext<WorkspaceBookmarkPayload>();
-
-    useBookmarkEvents({ applyFn: applyBookmark, saveFn: saveBookmark });
 };
