+++ src/Core/WorkSpace/src/Util/bookmarks/gardenBookmarks/gardenApiAtom.ts
@@ -1,4 +1,4 @@
 import { Atom } from '@dbeining/react-atom';
-import { GardenApi } from '../../../../../components/ParkView/Models/gardenApi';
+import { GardenApi } from '../../../../../../components/ParkView/Models/gardenApi';
 
 export const gardenApiAtom = Atom.of<GardenApi | null>(null);
