+++ src/Core/WorkSpace/src/Util/bookmarks/gardenBookmarks/generateGardenSnapshot.ts
@@ -1,4 +1,4 @@
-import { GardenApi } from '../../../../../components/ParkView/Models/gardenApi';
+import { GardenApi } from '../../../../../../components/ParkView/Models/gardenApi';
 import { GardenState } from './gardenStateSnapshotAtom';
 
 export function generateGardenSnapshot(api: GardenApi): GardenState {
