+++ src/Core/WorkSpace/src/Util/bookmarks/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './isWorkspaceBookmark';
