+++ src/Core/WorkSpace/src/Util/bookmarks/helpers/isWorkspaceBookmark.ts
@@ -1,0 +1,8 @@
+import { PowerBIBookmarkPayload } from '@equinor/lighthouse-powerbi';
+import { WorkspaceBookmarkPayload } from '../types';
+
+export const isWorkspaceBookmark = (
+    bookmark: Awaited<WorkspaceBookmarkPayload | PowerBIBookmarkPayload | never>
+): bookmark is WorkspaceBookmarkPayload => {
+    return (bookmark as WorkspaceBookmarkPayload)?.activeTab !== undefined;
+};
