+++ src/Core/WorkSpace/src/Util/bookmarks/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useWorkspaceBookmarks';
