+++ src/Core/WorkSpace/src/Util/bookmarks/types/gardenPayload.ts
@@ -1,0 +1,5 @@
+export type GardenPayload = {
+    customGroupByKeys?: Record<string, unknown>;
+    gardenKey?: string;
+    groupByKeys?: string[];
+};
