+++ src/Core/WorkSpace/src/Util/bookmarks/types/index.ts
@@ -1,0 +1,3 @@
+export type { WorkspaceBookmarkPayload } from './workspaceBookmarkPayload';
+export type { SaveBookmark, ApplyBookmark } from './saveAndApply';
+export type { GardenPayload } from './gardenPayload';
